fix: move the resolver to the services
