+++ webapp/src/main/java/io/github/microcks/web/RestInvocationProcessor.java
@@ -38,10 +38,10 @@ import io.github.microcks.util.dispatcher.ProxyFallbackSpecification;
 import io.github.microcks.util.el.EvaluableRequest;
 import io.github.microcks.util.script.JsScriptEngineBinder;
 import io.github.microcks.util.script.ScriptEngineBinder;
-
 import io.github.microcks.util.tracing.CommonEvents;
-import io.github.microcks.util.tracing.OpenTelemetryResolver;
+import io.github.microcks.service.OpenTelemetryResolverService;
 import io.github.microcks.util.tracing.TraceUtil;
+
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.api.trace.StatusCode;
@@ -95,7 +95,7 @@ public class RestInvocationProcessor {
    private final ProxyService proxyService;
 
    private ScriptEngine scriptEngine;
-   private final OpenTelemetryResolver opentelemetryResolver;
+   private final OpenTelemetryResolverService opentelemetryResolverService;
 
    @Value("${mocks.enable-invocation-stats}")
    private Boolean enableInvocationStats;
@@ -109,13 +109,13 @@ public class RestInvocationProcessor {
     */
    public RestInvocationProcessor(ServiceStateRepository serviceStateRepository, ResponseRepository responseRepository,
          ApplicationContext applicationContext, ProxyService proxyService,
-         OpenTelemetryResolver opentelemetryResolver) {
+         OpenTelemetryResolverService opentelemetryResolverService) {
       this.serviceStateRepository = serviceStateRepository;
       this.responseRepository = responseRepository;
       this.applicationContext = applicationContext;
       this.proxyService = proxyService;
       this.scriptEngine = new ScriptEngineManager().getEngineByExtension("groovy");
-      this.opentelemetryResolver = opentelemetryResolver;
+      this.opentelemetryResolverService = opentelemetryResolverService;
    }
 
    /**
@@ -278,7 +278,8 @@ public class RestInvocationProcessor {
       String dispatchCriteria = null;
       Map<String, Object> requestContext = null;
       // Create an INTERNAL child span explicitly because Spring AOP does not apply to private/self-invoked methods.
-      Tracer tracer = opentelemetryResolver.getOpenTelemetry().getTracer(RestInvocationProcessor.class.getName());
+      Tracer tracer = opentelemetryResolverService.getOpenTelemetry()
+            .getTracer(RestInvocationProcessor.class.getName());
       Span childSpan = tracer.spanBuilder("computeDispatchCriteria").setSpanKind(SpanKind.INTERNAL).startSpan();
       try (Scope ignored = childSpan.makeCurrent()) {
          TraceUtil.enableExplainTracing();
